requirements.txt
boto3==1.36.10
fastapi==0.115.8
httpx==0.28.1
motor==3.7.0
openai==1.60.2
passlib==1.7.4
pydantic==2.10.6
email-validator==2.2.0
pymongo==4.11
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.37
python-jose==3.3.0
python-multipart==0.0.20
argon2_cffi==23.1.0
pillow==11.1.0
uvicorn==0.34.0
slowapi==0.1.9
bleach==6.1.0
